update external website, and Android-WebMonkey userscript
add support for reading encoded DRM scheme and server URL in hash data
to most of the Chromecast senders;
each having an in-browser HTML5 video player,
and supporting DRM playback to a varying degree.

the limited testing that I performed on the HTML5 video players was disappointing;
none of them could play many of the video streams that I randomly selected from the wild.
YMMV.